requirements.txt
aniso8601==8.0.0
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
Pillow==7.0.0
pytz==2019.3
requests==2.22.0
six==1.14.0
soupsieve==1.9.5
SQLAlchemy==1.3.13
urllib3==1.25.8
Werkzeug==0.16.1
